update(almacen): add api barcode individual
