+++ templates/resources/resources_Almacen.py
@@ -21,7 +21,7 @@ from static.Models.api_inventory_models import (
     FileMovementsForm,
     file_movements_request_model,
     products_list_post_model,
-    ProductsListPostForm, movements_list_post_model, MovementsListPostForm,
+    ProductsListPostForm, movements_list_post_model, MovementsListPostForm, file_barcode_request_model, FileBarcodeForm,
 )
 from static.Models.api_movements_models import (
     movements_output_model,
@@ -42,7 +42,7 @@ from templates.resources.midleware.Functions_midleware_almacen import (
     upload_product_db_from_file,
     create_file_inventory,
     create_file_movements_amc,
-    insert_and_update_multiple_products_amc, insert_multiple_movements,
+    insert_and_update_multiple_products_amc, insert_multiple_movements, create_pdf_barcode,
 )
 from templates.controllers.product.p_and_s_controller import (
     delete_movement_db,
@@ -293,7 +293,14 @@ class DownloadMovementsFile(Resource):
 
 @ns.route("/inventory/file/download/barcode")
 class DownloadBarcodeFile(Resource):
-    @ns.expect(file_movements_request_model)
+    @ns.expect(file_barcode_request_model)
     def post(self):
-        filepath = "files/barcode.xlsx"
-        return send_file(filepath, as_attachment=True)
+        # noinspection PyUnresolvedReferences
+        validator = FileBarcodeForm.from_json(ns.payload)
+        if not validator.validate():
+            return {"data": validator.errors, "msg": "Error at structure"}, 400
+        data = validator.data
+        filepath, code = create_pdf_barcode(data)
+        return send_file(filepath, as_attachment=True) if code == 200 else (
+            {"data": filepath, "msg": "Error at creating file"}, 400
+        )
